Tom adding in dynamoDB and starting the calculation

+++ src/main/java/com/tom/autobetter/repository/dynamodb/AutobetterRepository.java
@@ -5,5 +5,5 @@ import org.socialsignin.spring.data.dynamodb.repository.EnableScan;
 import org.springframework.data.repository.CrudRepository;
 
 @EnableScan
-public interface AutobetterRepository extends CrudRepository<RaceDayEntity, Integer> {
+public interface AutobetterRepository extends CrudRepository<RaceDayEntity, Long> {
 }
